test: add three-tier test markers, fix nuts sign-compare, speed up slow tier
Split the pytest suite into cumulative fast / mid / slow tiers via markers so
the default `bb -t` inner-loop run stays fast (~8s):
- fast (default): unit/component tests
- mid (`bb -m`): full-pipeline .buda flow scripts (test_flow_scripts.py)
- slow (`bb -s`): SA/GA placement-optimizer storms

`bb` reworked: TEST_TIER replaces SLOW_TESTS (also fixing a latent bug where
`bb -s` built but never ran tests), added `-h/--help`, and unknown-arg errors.

Lowered SA `max_iter` 50k->12k in both test_optimize_demo_* tests after a
convergence sweep showed both seeds reach overlap=0 by ~8k iterations; cuts the
slow tier ~78s->~19s while keeping a ~1.5x margin.

Fixed a -Wsign-compare warning in nuts.cpp (size_t -> int to match
NUTSResult::num_overlaps).

Documented per-test runtimes and the tier rationale in
docs/internal/test_runtime_analysis.md; updated CLAUDE.md and pytest.ini.

+++ src/nuts.cpp
@@ -1837,7 +1837,7 @@ NUTSResult NUTSEngine::run(const std::vector<BundleWrapper>& bundles_in) {
                     jog_len = ts.span_hi - ts.span_lo;
             // Prefer: fewer overlaps, then the LONGER trunk (more jog room), then
             // the shorter jog.
-            const size_t ov = t.result.num_overlaps;
+            const int ov = t.result.num_overlaps;
             const bool better =
                 !applied ||
                 ov <  best_out.result.num_overlaps ||
